Refactor sidebar components: remove unused files and update imports for navigation buttons
