+++ app/components/sidebar/ui/index.ts
@@ -1,8 +1,8 @@
 // UI Components
 export { default as CampusButton } from "./campusButton";
-export { default as SidebarCloseButton } from "./sidebarCloseButton";
-export { default as SidebarNavigationButton } from "./sidebarNavigationButton";
-export { default as SidebarMobileNavigationButton } from "./sidebarMobileNavigationButton";
+export { default as SidebarCloseButton } from "./CloseButton";
+export { default as SidebarNavigationButton } from "./desktopNavigationButton";
+export { default as SidebarMobileNavigationButton } from "./mobileNavigationButton";
 export { default as SidebarToggleButton } from "./sidebarToggleButton";
 export { default as DragHandle } from "./dragHandle";
 export { default as SidebarHeader } from "./sidebarHeader";
